requirements.txt
aiohttp
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1 # To use the credentials with Azure SDK.
azure-mgmt-cognitiveservices # To interact with the Cognitive Services Management API.
#azure-search-documents==11.5.1
azure-search-documents==11.6.0b1
fastapi==0.115.2
fastapi-cli==0.0.5
httpx==0.27.2
ipykernel
msal==1.31.0 # Microsoft Authentication Library for Python.
msal-extensions==1.2.0
openai==1.43.1
passlib
pipreqs==0.5.0
pydantic==2.9.0
pydantic_core==2.23.2
python-multipart
pymongo==4.9.1
python-dotenv==1.0.1
python-jose[cryptography]
requests==2.32.3
starlette==0.40.0
starlette[sessions]
starlette[full]
uvicorn==0.32.0